backend/requirements.txt
# Web 框架和服务器
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# 图像处理
pillow==10.1.0

# HTTP 客户端
requests==2.31.0
httpx==0.25.1

# 工具库
python-dateutil==2.8.2
python-dotenv==1.0.0
